hooks/oclint now returns 1 if there is an error

+++ tests/hooks_test.py
@@ -29,24 +29,34 @@ class TestCLinters:
         cls.testfiles = [os.path.abspath(filename) for filename in testfiles]
 
     def test_clang_format(self):
-        """Test clang format using google style, printing to stdout."""
-        expected = "int main() {\n  int i;\n  return 10;\n}"
+        """Use google style, printing a format diff to stdout."""
+        expected = r"""1c1,4
+< int main() { int i; return 10; }
+\ No newline at end of file
+---
+> int main() {
+>   int i;
+>   return 10;
+> }
+\ No newline at end of file
+"""
         for filename in self.testfiles:
-            actual = sp.check_output(
-                ["hooks/clang-format", filename], text=True
+            _pipe = sp.Popen(
+                ["hooks/clang-format", filename],
+                text=True,
+                stdout=sp.PIPE,
+                stderr=sp.STDOUT,
             )
+            actual = _pipe.communicate()[0]
             assert actual == expected
 
     def test_clang_tidy(self):
         """Test clang tidy using all checks."""
-        clang_tidy_expected = """\
-2 warnings generated.
-{0}:1:28: error: 10 is a magic number; \
-consider replacing it with a named constant \
-[cppcoreguidelines-avoid-magic-numbers,-warnings-as-errors]
+        clang_tidy_expected = r"""2 warnings generated.
+{0}:1:28: error: 10 is a magic number; consider replacing it with a named constant [cppcoreguidelines-avoid-magic-numbers,-warnings-as-errors]
 int main() {{ int i; return 10; }}
                            ^
-"""
+"""  # noqa: E501
         cmds = [
             "./hooks/clang-tidy",
             "-quiet",
@@ -63,19 +73,17 @@ int main() {{ int i; return 10; }}
     @pytest.mark.slow
     def test_oclint(self):
         """Test oclint with major analyses turned on."""
-        oclint_expected_template = """\
-
+        oclint_expected_template = r"""
 
 OCLint Report
 
 Summary: TotalFiles=1 FilesWithViolations=1 P1=0 P2=0 P3=2
 
-{0}:1:14: short variable name [naming|P3] Length of variable name `i` is 1, \
-which is shorter than the threshold of 3
+{0}:1:14: short variable name [naming|P3] Length of variable name `i` is 1, which is shorter than the threshold of 3
 {0}:1:14: unused local variable [unused|P3] The local variable 'i' is unused.
 
 [OCLint (http://oclint.org) v0.13]
-"""
+"""  # noqa: E501
         cmds = [
             "./hooks/oclint",
             "-enable-global-analysis",
@@ -97,8 +105,10 @@ which is shorter than the threshold of 3
             Text output of function
         """
         combined_cmds = cmds + [filename]
-        _pipe = sp.Popen(combined_cmds, stderr=sp.STDOUT, stdout=sp.PIPE)
-        return _pipe.communicate()[0].decode("utf-8")
+        _pipe = sp.Popen(
+            combined_cmds, stderr=sp.STDOUT, stdout=sp.PIPE, text=True
+        )
+        return _pipe.communicate()[0]
 
     @classmethod
     def teardown_class(cls):
@@ -107,4 +117,6 @@ which is shorter than the threshold of 3
         if os.path.exists("test.plist"):
             os.remove("test.plist")
         for filename in cls.testfiles:
-            os.remove(filename)
+            filename = os.path.abspath(filename)
+            if os.path.exists(filename):
+                os.remove(filename)
